+++ CHANGELOG.md
@@ -1,11 +1,11 @@
-﻿# Changelog
+# Changelog
 
 All notable changes to this project will be documented in this file.
 
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## 3.1.0 - 2026-06-18
 
 * Added: New alias pw for New-Password and allow multiple output types (New-Password)
 * Changed: System audit optimization: Better sort and output formatting (Get-SystemAudit)